USRE-87 Merge MetaData test to one test
