+++ src/test/java/com/ossez/usreio/tests/client/RetsMetadataTest.java
@@ -1,16 +1,21 @@
 package com.ossez.usreio.tests.client;
 
+import java.io.InputStream;
 import java.net.MalformedURLException;
+import java.util.Properties;
 
+import com.ossez.usreio.client.retsapi.RETSConnection;
+import com.ossez.usreio.client.retsapi.RETSGetMetadataTransaction;
 import com.ossez.usreio.tests.common.metadata.types.MClass;
 import com.ossez.usreio.tests.common.metadata.types.MResource;
 import com.ossez.usreio.tests.common.metadata.types.MSystem;
 import com.ossez.usreio.client.*;
+import org.junit.jupiter.api.Test;
 
 /**
  * Simple Example performing a GetMetadata and iterating of the results
  */
-public class RetsGetMetadataExample {
+public class RetsMetadataTest {
 
     public static void main(String[] args) throws MalformedURLException {
 
@@ -64,5 +69,41 @@ public class RetsGetMetadataExample {
                 }
             }
         }
+    }
+
+    /**
+     * Do RetsServerConnection Test
+     */
+    @Test
+    public void testStaticVariableChange() {
+
+        //        BasicConfigurator.configure();
+
+        RETSConnection rc = new RETSConnection();
+//        RETSLoginTransaction trans = new RETSLoginTransaction();
+        RETSGetMetadataTransaction trans = new RETSGetMetadataTransaction();
+
+
+        try {
+            Properties props = new Properties();
+            ClassLoader loader = Thread.currentThread().getContextClassLoader();
+            InputStream inputStream = loader.getResourceAsStream("rets.properties");
+
+            props.load(inputStream);
+
+            // Add the optional request parameters if they exist, are non-null and non-zero-length
+            // rc.setRequestHeaderField("Authorization", (String)props.get("login.AUTHORIZATION"));
+            rc.setServerUrl((String) props.getProperty("rets_server"));
+//            trans.setUrl((String) props.getProperty("rets_server"));
+//            trans.setUsername((String) props.getProperty("rets_username"));
+//            trans.setPassword((String) props.getProperty("rets_password"));
+        } catch (Exception e) {
+            e.printStackTrace();
+        }
+        rc.execute(trans);
+//        rc.execute(transaction);
+
+//        transaction.getVersion();
+
     }
 }
