+++ src/test/java/com/ossez/usreio/tests/client/RetsSessionTest.java
@@ -4,12 +4,13 @@ import com.ossez.usreio.client.RetsException;
 import com.ossez.usreio.client.RetsSession;
 import com.ossez.usreio.client.RetsVersion;
 import com.ossez.usreio.util.SessionUtils;
+import org.apache.commons.lang3.ObjectUtils;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.TestInstance;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import static org.junit.jupiter.api.Assertions.assertNull;
+import static org.junit.jupiter.api.Assertions.assertNotNull;
 
 /**
  * Test for RETS session
@@ -22,14 +23,15 @@ public class RetsSessionTest extends RetsTestCase {
 
 
     /**
-     * Test Login
+     * Test Login should return SessionID from server
      */
     @Test
     public void testLogin() {
+        logger.debug("Test Rets Session Login by URL: [{}]", retsLoginUrl);
 
         try {
             RetsSession session = SessionUtils.retsLogin(retsLoginUrl, retsUsername, retsPassword, RetsVersion.RETS_1_7_2);
-            assertNull(session.getSessionId());
+            assertNotNull(session.getSessionId());
         } catch (RetsException ex) {
             logger.debug("Session Login Error", ex);
         }
@@ -40,28 +42,28 @@ public class RetsSessionTest extends RetsTestCase {
     /**
      * TEST Logout
      */
-    @org.junit.jupiter.api.Test
+    @Test
     public void testLogout() {
-        logger.debug("RETS Session Logout URL: [{}]", retsLoginUrl);
+        logger.debug("RETS Session Login URL: [{}]", retsLoginUrl);
         RetsSession session = null;
 
         try {
             session = SessionUtils.retsLogin(retsLoginUrl, retsUsername, retsPassword, RetsVersion.RETS_1_7_2);
-//            Assert.assertNotNull(session.getSessionId());
+            assertNotNull(session.getSessionId());
         } catch (RetsException ex) {
             logger.debug("Session Login Error", ex);
         }
 
 
-        if (session != null) {
+        // If Session is not Empty then Logout
+        if (ObjectUtils.isNotEmpty(session)) {
             try {
                 session.logout();
             } catch (RetsException e) {
-                e.printStackTrace();
+                logger.error("Logout Error: ", e);
             }
         }
 
     }
-
 
 }
